# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Reimbursement Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: Get expenses (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: userId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseWithDetails"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Submit a new expense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Expense submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseWithDetails"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}/approve:
    post:
      operationId: approveExpense
      tags: [approvals]
      summary: Approve or reject an expense
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalActionRequest"
      responses:
        "200":
          description: Approval action taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseWithDetails"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized to approve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}/override:
    post:
      operationId: overrideExpense
      tags: [approvals]
      summary: Admin override approval
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverrideRequest"
      responses:
        "200":
          description: Override applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseWithDetails"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /upload/receipt:
    post:
      operationId: uploadReceipt
      tags: [expenses]
      summary: Upload receipt image (base64)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadReceiptRequest"
      responses:
        "200":
          description: Receipt uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadReceiptResponse"
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [expenses]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        managerId:
          type: ["integer", "null"]
        companyId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - companyId
        - createdAt
...
    ExpenseWithDetails:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        convertedAmount:
          type: ["number", "null"]
        category:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        receiptUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        currentStep:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalWithApprover"
      required:
        - id
        - userId
        - amount
        - currency
        - category
        - description
        - date
        - status
        - createdAt
        - user
        - approvals
    CreateExpenseRequest:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        category:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        receiptUrl:
          type: ["string", "null"]
      required:
        - amount
        - currency
        - category
        - description
        - date
    ApprovalActionRequest:
      type: object
      properties:
        action:
          type: string
          enum: [approved, rejected]
        comments:
          type: string
      required:
        - action
    OverrideRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        comments:
          type: string
      required:
        - status
...
    ApprovalWithApprover:
      type: object
      properties:
        id:
          type: integer
        expenseId:
          type: integer
        approverId:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
        comments:
          type: ["string", "null"]
        stepOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        approver:
          $ref: "#/components/schemas/User"
      required:
        - id
        - expenseId
        - approverId
        - status
        - stepOrder
        - createdAt
        - updatedAt
        - approver
...
    UploadReceiptRequest:
      type: object
      properties:
        imageBase64:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
      required:
        - imageBase64
        - fileName
    UploadReceiptResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    DashboardStats:
      type: object
      properties:
        totalExpenses:
          type: integer
        pendingExpenses:
          type: integer
        approvedExpenses:
          type: integer
        rejectedExpenses:
          type: integer
        totalAmount:
          type: number
        pendingAmount:
          type: number
        approvedAmount:
          type: number
        teamExpenses:
          type: integer
        recentExpenses:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseWithDetails"
      required:
        - totalExpenses
        - pendingExpenses
        - approvedExpenses
        - rejectedExpenses
        - totalAmount
        - pendingAmount
        - approvedAmount
        - teamExpenses
        - recentExpenses