# lib/api-zod/src/generated/types/approvalWithApprover.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reimbursement Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalWithApproverStatus } from "./approvalWithApproverStatus";
import type { User } from "./user";

export interface ApprovalWithApprover {
  id: number;
  expenseId: number;
  approverId: number;
  status: ApprovalWithApproverStatus;
  /** @nullable */
  comments?: string | null;
  stepOrder: number;
  createdAt: Date;
  updatedAt: Date;
  approver: User;
}